+++ backend/bots/examples/satoshi_dice/impl/src/jobs/process_pending_actions.rs
@@ -98,7 +98,7 @@ async fn process_action(action: Action) {
                                     recipient: user_id,
                                     transfer: CryptoTransaction::Completed(CompletedCryptoTransaction::SNS(
                                         sns::CompletedCryptoTransaction {
-                                            ledger: Cryptocurrency::CKBTC.ledger_canister_id(),
+                                            ledger: Cryptocurrency::CKBTC.ledger_canister_id().unwrap(),
                                             token: Cryptocurrency::CKBTC,
                                             amount: Tokens::from_e8s(amount),
                                             fee: Tokens::from_e8s(10),