+++ backend/canisters/user_index/impl/src/jobs/make_pending_payments.rs
@@ -66,7 +66,8 @@ async fn make_payment(pending_payment: &PendingPayment) -> Result<BlockIndex, bo
         amount: pending_payment.amount.into(),
     };
 
-    match icrc1_ledger_canister_c2c_client::icrc1_transfer(pending_payment.currency.ledger_canister_id(), &args).await {
+    match icrc1_ledger_canister_c2c_client::icrc1_transfer(pending_payment.currency.ledger_canister_id().unwrap(), &args).await
+    {
         Ok(Ok(block_index)) => Ok(block_index.0.try_into().unwrap()),
         Ok(Err(transfer_error)) => {
             error!(?transfer_error, ?args, "Transfer failed");